crispy-bot/pyproject.toml
[project]
name = "crispy-bot"
version = "0.1.0"
description = "crispy-bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.5.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-console>=0.10.0",
    "nonebot-plugin-docs>=2.5.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-abbrreply>=1.1.3",
    "litellm>=1.83.8",
    "httpx>=0.28.1",
]

[tool.nonebot]
plugin_dirs = []
builtin_plugins = ["echo"]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" }
]
nonebot-adapter-console = [
    { name = "Console", module_name = "nonebot.adapters.console" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-docs = ["nonebot_plugin_docs"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-abbrreply = ["nonebot_plugin_abbrreply"]
